Update constructors to return struct references instead of interfaces
